Test: TrackServiceTest 구현

+++ src/main/java/seoul/AutoEveryDay/service/TrackService.java
@@ -25,28 +25,23 @@ public class TrackService {
     private final TrackRepository trackRepository;
     private final DriveHistoryRepository driveHistoryRepository;
 
-    private void validateTestHistory(DriveHistoryDto testHistory, User user) {
+    private void validateDriveHistory(DriveHistoryDto testHistory, User user) {
         if (testHistory.getDate().isBefore(LocalDate.now())) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "예약은 오늘 이후로만 가능합니다.");
         }
         if (driveHistoryRepository.findByTrackIdAndDate(user.getId(), testHistory.getDate()).isPresent()) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이미 예약한 날짜입니다.");
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이미 예약된 날짜입니다.");
         }
     }
-    public Track getTestTrack(String name) {
-        return trackRepository.findByName(name).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "테스트 센터를 찾을 수 없습니다.")
-        );
-    }
-    public Track getTestTrack(Long id) {
+    private Track getTrack(Long id) {
         return trackRepository.findById(id).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "테스트 센터를 찾을 수 없습니다.")
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "트랙을 찾을 수 없습니다.")
         );
     }
 
     public TrackDto createTestTrack(TrackDto trackDto) {
         if (trackRepository.existsByName(trackDto.getName())) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이미 존재하는 테스트 센터 이름입니다.");
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이미 존재하는 트랙 입니다.");
         }
         Track track = Track.builder()
                 .name(trackDto.getName())
@@ -58,14 +53,15 @@ public class TrackService {
             log.error("테스트 센터 저장 실패", e);
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "테스트 센터 저장 실패");
         }
+        trackDto.setId(track.getId());
         return trackDto;
     }
     public TrackDto editTestTrack(TrackDto trackDto) {
-        Track track = getTestTrack(trackDto.getName());
+        Track track = getTrack(trackDto.getId());
         trackRepository.findByName(trackDto.getName()).ifPresent(
                 (t) -> {
                     if (!t.getId().equals(track.getId())) {
-                        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이미 존재하는 테스트 센터 이름입니다.");
+                        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "이미 존재하는 트랙 입니다.");
                     }
                 }
         );
@@ -73,8 +69,11 @@ public class TrackService {
         track.setDescription(trackDto.getDescription());
         return trackDto;
     }
-    public TrackDto deleteTestTrack(String name) {
-        Track track = getTestTrack(name);
+    public TrackDto deleteTestTrack(Long id) {
+        Track track = getTrack(id);
+        if (driveHistoryRepository.existsByTrackId(track.getId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "예약된 트랙은 삭제할 수 없습니다.");
+        }
         try {
             trackRepository.delete(track);
         } catch (Exception e) {
@@ -82,6 +81,7 @@ public class TrackService {
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "테스트 센터 삭제 실패");
         }
         return TrackDto.builder()
+                .id(track.getId())
                 .name(track.getName())
                 .description(track.getDescription())
                 .build();
@@ -89,6 +89,7 @@ public class TrackService {
     public List<TrackDto> getAllTestTrack() {
         return trackRepository.findAll().stream()
                 .map((t) -> TrackDto.builder()
+                        .id(t.getId())
                         .name(t.getName())
                         .description(t.getDescription())
                         .build()
@@ -96,9 +97,9 @@ public class TrackService {
     }
 
     // 여기부터 예약 관련
-    public DriveHistoryDto createTestHistory(DriveHistoryDto driveHistoryDto, User user) {
-        validateTestHistory(driveHistoryDto, user);
-        Track track = getTestTrack(driveHistoryDto.getTrackId());
+    public DriveHistoryDto createDriveHistory(DriveHistoryDto driveHistoryDto, User user) {
+        validateDriveHistory(driveHistoryDto, user);
+        Track track = getTrack(driveHistoryDto.getTrackId());
         DriveHistory driveHistory = DriveHistory.builder()
                 .user(user)
                 .track(track)
@@ -110,14 +111,16 @@ public class TrackService {
             log.error("예약 저장 실패", e);
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "예약 저장 실패");
         }
+        driveHistory.setId(driveHistory.getId());
         return driveHistoryDto;
     }
 
-    public DriveHistoryDto deleteTestHistory(DriveHistoryDto driveHistoryDto, User user) {
-        Track track = getTestTrack(driveHistoryDto.getTrackId());
+    public DriveHistoryDto deleteDriveHistory(DriveHistoryDto driveHistoryDto, User user) {
+        Track track = getTrack(driveHistoryDto.getTrackId());
         DriveHistory driveHistory = driveHistoryRepository.findByUserIdAndTrackIdAndDate(user.getId(), track.getId(), driveHistoryDto.getDate()).orElseThrow(
                 () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "예약을 찾을 수 없습니다.")
         );
+        driveHistoryDto.setId(driveHistory.getId());
         try {
             driveHistoryRepository.delete(driveHistory);
         } catch (Exception e) {
